+++ samples/star/src/iosMain/kotlin/com/slack/circuit/star/StarUiViewController.kt
@@ -4,9 +4,18 @@ package com.slack.circuit.star
 
 import androidx.compose.ui.window.ComposeUIViewController
 import com.slack.circuit.foundation.CircuitCompositionLocals
+import com.slack.circuit.retained.CircuitRetainedSettings
+import com.slack.circuit.retained.ExperimentalCircuitRetainedApi
+import com.slack.circuit.retained.RetainedValuesStoreProvider
 import com.slack.circuit.star.di.AppGraph
 import platform.UIKit.UIViewController
 
-fun makeUiViewController(graph: AppGraph): UIViewController = ComposeUIViewController {
-  CircuitCompositionLocals(graph.circuit) { StarCircuitApp() }
+@OptIn(ExperimentalCircuitRetainedApi::class)
+fun makeUiViewController(graph: AppGraph): UIViewController {
+  CircuitRetainedSettings.useFirstParty = true
+  return ComposeUIViewController {
+    RetainedValuesStoreProvider {
+      CircuitCompositionLocals(graph.circuit) { StarCircuitApp() }
+    }
+  }
 }