+++ samples/star/src/jvmMain/kotlin/com/slack/circuit/star/main.kt
@@ -23,11 +23,16 @@ import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
 import coil3.SingletonImageLoader
 import com.slack.circuit.foundation.CircuitCompositionLocals
+import com.slack.circuit.retained.CircuitRetainedSettings
+import com.slack.circuit.retained.ExperimentalCircuitRetainedApi
+import com.slack.circuit.retained.RetainedValuesStoreProvider
 import com.slack.circuit.star.di.AppGraph
 import java.awt.Desktop
 import java.net.URI
 
+@OptIn(ExperimentalCircuitRetainedApi::class)
 fun main() {
+  CircuitRetainedSettings.useFirstParty = true
   val appGraph = AppGraph.create()
   SingletonImageLoader.setSafe { appGraph.imageLoader }
   application {
@@ -80,7 +85,7 @@ fun main() {
               }
             }
           }
-          StarCircuitApp(state = state)
+          RetainedValuesStoreProvider { StarCircuitApp(state = state) }
         }
       }
     }